+++ src/senaite/databox/behaviors/databox.py
@@ -22,6 +22,7 @@ import ast
 from copy import copy
 from datetime import datetime
 
+import transaction
 from bika.lims import api
 from DateTime import DateTime
 from dateutil import parser
@@ -230,7 +231,10 @@ class DataBox(object):
         temp_folder = portal_factory._getTempFolder(TMP_FOLDER_KEY)
         if portal_type in temp_folder:
             return temp_folder[portal_type]
+        # reduce conflict errors
+        portal_factory._p_jar.sync()
         temp_folder.invokeFactory(portal_type, id=portal_type)
+        transaction.commit()
         return temp_folder[portal_type]
 
     def get_query_catalog(self, default="portal_catalog"):