+++ docs/Changelog.rst
@@ -1,6 +1,7 @@
 1.1.0 (unreleased)
 ------------------
 
+- #14 Reduce conflict errors when creating temporary objects
 - #12 Remove orphane reference columns
 - #11 Fix error when the column "SampleType" was selected #11
 - #10 Fix Traceback for Dexterity Types