+++ docs/Changelog.rst
@@ -1,6 +1,7 @@
 1.1.0 (unreleased)
 ------------------
 
+- #16 Added upgrade step handler
 - #14 Reduce conflict errors when creating temporary objects
 - #12 Remove orphane reference columns
 - #11 Fix error when the column "SampleType" was selected #11