+++ docs/Changelog.rst
@@ -1,4 +1,4 @@
-1.1.0 (unreleased)
+1.1.0 (2022-01-05)
 ------------------
 
 - #22 Support datetime objects in date converter
@@ -11,7 +11,6 @@
 - #12 Remove orphane reference columns
 - #11 Fix error when the column "SampleType" was selected #11
 - #10 Fix Traceback for Dexterity Types
-
 
 
 1.0.0 (2020-09-22)